feat: promote experimental finder APIs into the main TCT module
Move the former TCT.experimental wrappers (pathfinder,
neighborhood_finder, resource cache, result dataclasses) into
TCT/TCT.py, replacing the legacy Path_finder and Neighborhood_finder
entry points. Delete the unused low-level TCT_pathfinder.pathfinder
and rename tests/test_experimental.py to tests/test_finder_api.py.
