--- TCT/experimental.py
@@ -1,598 +1,0 @@
-"""
-Experimental developer-friendly APIs for common Translator finder patterns.
-
-This module provides higher-level wrappers around the existing pathfinder and
-neighborhood finder code. The wrappers resolve human-readable names, normalize
-CURIEs, load Translator resources lazily, and return a small result object with
-the most useful output fields surfaced directly.
-"""
-
-from dataclasses import dataclass
-from typing import Any, Optional, TypeAlias, Union
-
-import pandas as pd
-
-from . import name_resolver, node_normalizer, translator_query
-from .TCT import sele_predicates_API
-from .TCT_neighborhood_finder import (
-    parse_results_for_neighborhood_finder,
-    parse_results_for_neighborhood_finder_multiple_inputs,
-)
-from .TCT_pathfinder import parse_results_for_pathfinder
-
-
-NodeInput: TypeAlias = str
-CategoryInput: TypeAlias = str
-CategoryList: TypeAlias = list[CategoryInput]
-
-
-@dataclass(frozen=True)
-class ResolvedNode:
-    """Resolved node metadata used by the experimental finder APIs."""
-
-    input_value: str
-    curie: str
-    label: Optional[str]
-    categories: list[str]
-
-
-@dataclass
-class TranslatorResources:
-    """Translator API metadata required by the legacy query functions."""
-
-    api_names: dict[str, str]
-    meta_kg: pd.DataFrame
-    api_predicates: dict[str, list[str]]
-
-
-@dataclass
-class FinderResult:
-    """Convenience wrapper around a parsed TRAPI-style finder response."""
-
-    query: dict[str, Any]
-    knowledge_graph: dict[str, Any]
-    results: list[dict[str, Any]]
-    auxiliary_graphs: dict[str, Any]
-    resolved_nodes: dict[str, ResolvedNode]
-    raw: dict[str, Any]
-
-    def to_dict(self) -> dict[str, Any]:
-        """
-        Return the raw parsed TRAPI-style output dictionary.
-
-        Returns
-        -------
-        dict
-            Full parsed output generated by the existing finder parser.
-
-        Examples
-        --------
-        >>> result = FinderResult({}, {}, [], {}, {}, {})
-        >>> result.to_dict()
-        {}
-        """
-        return self.raw
-
-
-_DEFAULT_TRANSLATOR_RESOURCES: Optional[TranslatorResources] = None
-
-
-def get_translator_resources(*, refresh: bool = False) -> TranslatorResources:
-    """
-    Return cached Translator API metadata, loading it on first use.
-
-    Parameters
-    ----------
-    refresh : bool
-        If true, refetch SmartAPI/MetaKG data even when the singleton is
-        already populated.
-
-    Returns
-    -------
-    TranslatorResources
-        API names, MetaKG dataframe, and API predicate mapping used by the
-        legacy query functions.
-
-    Examples
-    --------
-    >>> resources = get_translator_resources()
-    >>> paths = pathfinder("asthma", "albuterol", ["Gene"], resources=resources)
-    """
-    global _DEFAULT_TRANSLATOR_RESOURCES
-    if refresh or _DEFAULT_TRANSLATOR_RESOURCES is None:
-        api_names, meta_kg, api_predicates = (
-            translator_query.get_translator_API_predicates()
-        )
-        _DEFAULT_TRANSLATOR_RESOURCES = TranslatorResources(
-            api_names=api_names,
-            meta_kg=meta_kg,
-            api_predicates=api_predicates,
-        )
-    return _DEFAULT_TRANSLATOR_RESOURCES
-
-
-def clear_translator_resource_cache() -> None:
-    """
-    Clear the in-memory Translator resource singleton.
-
-    Examples
-    --------
-    >>> clear_translator_resource_cache()
-    >>> resources = get_translator_resources()  # refetches
-    """
-    global _DEFAULT_TRANSLATOR_RESOURCES
-    _DEFAULT_TRANSLATOR_RESOURCES = None
-
-
-def pathfinder(
-    start: NodeInput,
-    end: NodeInput,
-    intermediate_categories: CategoryList,
-    *,
-    start_categories: Optional[CategoryList] = None,
-    end_categories: Optional[CategoryList] = None,
-    api_names: Optional[dict[str, str]] = None,
-    meta_kg: Optional[pd.DataFrame] = None,
-    api_predicates: Optional[dict[str, list[str]]] = None,
-    resources: Optional[TranslatorResources] = None,
-    scoring_method: str = "infores",
-    name_resolver_kwargs: Optional[dict[str, Any]] = None,
-    node_normalizer_kwargs: Optional[dict[str, Any]] = None,
-) -> FinderResult:
-    """
-    Find paths between two biomedical concepts using Translator KPs.
-
-    Parameters
-    ----------
-    start : str
-        Start node as either a CURIE (for example, ``"MONDO:0004979"``) or a
-        human-readable string (for example, ``"asthma"``).
-    end : str
-        End node as either a CURIE or human-readable string.
-    intermediate_categories : list[str]
-        Allowed categories for intermediate path nodes. Values may be short
-        names like ``"Gene"`` or full Biolink names like ``"biolink:Gene"``.
-    start_categories : list[str], optional
-        Category override for the start node. If omitted, categories are
-        inferred from Node Normalizer.
-    end_categories : list[str], optional
-        Category override for the end node. If omitted, categories are inferred
-        from Node Normalizer.
-    resources : TranslatorResources, optional
-        Preloaded Translator resources. If omitted, the module-level singleton
-        is loaded on first use and reused.
-    api_names, meta_kg, api_predicates : optional
-        Advanced partial overrides for the Translator resources used by the
-        legacy pathfinder implementation.
-    scoring_method : str
-        Scoring method passed to the legacy parser. Current values are
-        ``"infores"`` and ``"edges"``.
-    name_resolver_kwargs : dict, optional
-        Extra keyword arguments for ``name_resolver.lookup``.
-    node_normalizer_kwargs : dict, optional
-        Extra keyword arguments for ``node_normalizer.get_normalized_nodes``.
-
-    Returns
-    -------
-    FinderResult
-        Convenience wrapper containing resolved input nodes, the parsed
-        knowledge graph, results, auxiliary graphs, and the raw TRAPI-style
-        output dictionary.
-
-    Examples
-    --------
-    >>> from TCT.experimental import pathfinder
-    >>> result = pathfinder("asthma", "albuterol", ["Gene"])
-    >>> result.resolved_nodes["start"].curie
-    'MONDO:0004979'
-    """
-    start_node = _resolve_node(
-        start,
-        name_resolver_kwargs=name_resolver_kwargs,
-        node_normalizer_kwargs=node_normalizer_kwargs,
-    )
-    end_node = _resolve_node(
-        end,
-        name_resolver_kwargs=name_resolver_kwargs,
-        node_normalizer_kwargs=node_normalizer_kwargs,
-    )
-    intermediate_categories = _normalize_categories(intermediate_categories) or []
-    start_categories = _normalize_categories(start_categories) or start_node.categories
-    end_categories = _normalize_categories(end_categories) or end_node.categories
-    resolved_resources = _get_resources(
-        resources=resources,
-        api_names=api_names,
-        meta_kg=meta_kg,
-        api_predicates=api_predicates,
-    )
-
-    predicates1, apis1, _ = sele_predicates_API(
-        start_categories,
-        intermediate_categories,
-        resolved_resources.meta_kg,
-        resolved_resources.api_names,
-    )
-    predicates2, apis2, _ = sele_predicates_API(
-        intermediate_categories,
-        end_categories,
-        resolved_resources.meta_kg,
-        resolved_resources.api_names,
-    )
-    query1 = translator_query.format_query_json(
-        [start_node.curie],
-        [],
-        start_categories,
-        intermediate_categories,
-        predicates1,
-    )
-    query2 = translator_query.format_query_json(
-        [],
-        [end_node.curie],
-        intermediate_categories,
-        end_categories,
-        predicates2,
-    )
-    result1 = translator_query.parallel_api_query(
-        query_json=query1,
-        select_APIs=apis1,
-        APInames=resolved_resources.api_names,
-        API_predicates=resolved_resources.api_predicates,
-        max_workers=max(1, len(apis1)),
-    )
-    result2 = translator_query.parallel_api_query(
-        query_json=query2,
-        select_APIs=apis2,
-        APInames=resolved_resources.api_names,
-        API_predicates=resolved_resources.api_predicates,
-        max_workers=max(1, len(apis2)),
-    )
-    raw_output = parse_results_for_pathfinder(
-        start_node.curie,
-        end_node.curie,
-        result1,
-        result2,
-        start_node_categories=start_categories,
-        end_node_categories=end_categories,
-        scoring_method=scoring_method,
-        get_node_info=True,
-    )
-    return _build_finder_result(
-        raw_output,
-        resolved_nodes={"start": start_node, "end": end_node},
-    )
-
-
-def neighborhood_finder(
-    node: Union[NodeInput, list[NodeInput]],
-    neighbor_categories: CategoryList,
-    *,
-    node_categories: Optional[CategoryList] = None,
-    api_names: Optional[dict[str, str]] = None,
-    meta_kg: Optional[pd.DataFrame] = None,
-    api_predicates: Optional[dict[str, list[str]]] = None,
-    resources: Optional[TranslatorResources] = None,
-    predicates_subset: Optional[list[str]] = None,
-    attribute_constraints: Optional[list[dict[str, Any]]] = None,
-    name_resolver_kwargs: Optional[dict[str, Any]] = None,
-    node_normalizer_kwargs: Optional[dict[str, Any]] = None,
-) -> FinderResult:
-    """
-    Find one-hop neighbors for one or more biomedical concepts.
-
-    Parameters
-    ----------
-    node : str or list[str]
-        Source node or nodes. Each value may be a CURIE or human-readable
-        string. Human-readable strings are resolved with Name Resolver and then
-        normalized with Node Normalizer.
-    neighbor_categories : list[str]
-        Desired neighbor categories. Values may be short names like ``"Drug"``
-        or full Biolink names like ``"biolink:Drug"``.
-    node_categories : list[str], optional
-        Category override for source nodes. If omitted, categories are inferred
-        from the first normalized source node.
-    resources : TranslatorResources, optional
-        Preloaded Translator resources. If omitted, the module-level singleton
-        is loaded on first use and reused.
-    api_names, meta_kg, api_predicates : optional
-        Advanced partial overrides for the Translator resources used by the
-        legacy neighborhood implementation.
-    predicates_subset : list[str], optional
-        Optional predicate filter applied after MetaKG predicate selection.
-    attribute_constraints : list[dict], optional
-        TRAPI attribute constraints passed through to query construction.
-    name_resolver_kwargs : dict, optional
-        Extra keyword arguments for ``name_resolver.lookup``.
-    node_normalizer_kwargs : dict, optional
-        Extra keyword arguments for ``node_normalizer.get_normalized_nodes``.
-
-    Returns
-    -------
-    FinderResult
-        Convenience wrapper containing resolved input nodes, parsed neighborhood
-        knowledge graph, results, auxiliary graphs, and raw TRAPI-style output.
-
-    Examples
-    --------
-    >>> from TCT.experimental import neighborhood_finder
-    >>> result = neighborhood_finder("asthma", ["SmallMolecule", "Drug"])
-    >>> result.knowledge_graph["nodes"]
-    {...}
-    """
-    resolved_nodes = _resolve_nodes(
-        node,
-        name_resolver_kwargs=name_resolver_kwargs,
-        node_normalizer_kwargs=node_normalizer_kwargs,
-    )
-    source_categories = (
-        _normalize_categories(node_categories) or resolved_nodes[0].categories
-    )
-    neighbor_categories = _normalize_categories(neighbor_categories) or []
-    resolved_resources = _get_resources(
-        resources=resources,
-        api_names=api_names,
-        meta_kg=meta_kg,
-        api_predicates=api_predicates,
-    )
-
-    predicates, apis, _ = sele_predicates_API(
-        source_categories,
-        neighbor_categories,
-        resolved_resources.meta_kg,
-        resolved_resources.api_names,
-    )
-    if predicates_subset is not None:
-        predicates = list(set(predicates).intersection(predicates_subset))
-    if len(predicates) == 0:
-        predicates = ["biolink:related_to"]
-
-    input_curies = [resolved_node.curie for resolved_node in resolved_nodes]
-    query = translator_query.format_query_json(
-        subject_ids=input_curies,
-        object_ids=None,
-        subject_categories=None,
-        object_categories=neighbor_categories,
-        predicates=predicates,
-        attribute_constraints=attribute_constraints,
-    )
-    raw_edges = translator_query.parallel_api_query(
-        query_json=query,
-        select_APIs=apis,
-        APInames=resolved_resources.api_names,
-        API_predicates=resolved_resources.api_predicates,
-        max_workers=max(1, len(apis)),
-    )
-    if isinstance(node, str):
-        raw_output = parse_results_for_neighborhood_finder(
-            input_curies[0],
-            raw_edges,
-            source_categories,
-            neighbor_categories,
-        )
-        result_nodes = {"node": resolved_nodes[0]}
-    else:
-        raw_output = parse_results_for_neighborhood_finder_multiple_inputs(
-            input_curies,
-            raw_edges,
-            source_categories,
-            neighbor_categories,
-        )
-        result_nodes = {
-            f"node_{index}": resolved_node
-            for index, resolved_node in enumerate(resolved_nodes)
-        }
-    return _build_finder_result(raw_output, resolved_nodes=result_nodes)
-
-
-def _normalize_category(category: str) -> str:
-    """
-    Convert a category into a Biolink-prefixed category string.
-
-    Parameters
-    ----------
-    category : str
-        Category name, with or without the ``biolink:`` prefix.
-
-    Returns
-    -------
-    str
-        Biolink-prefixed category.
-
-    Examples
-    --------
-    >>> _normalize_category("Gene")
-    'biolink:Gene'
-    """
-    category = category.strip()
-    if category.startswith("biolink:"):
-        return category
-    return f"biolink:{category}"
-
-
-def _normalize_categories(categories: Optional[CategoryList]) -> Optional[list[str]]:
-    """
-    Normalize a list of category strings.
-
-    Parameters
-    ----------
-    categories : list[str], optional
-        Category names, with or without ``biolink:`` prefixes.
-
-    Returns
-    -------
-    list[str] or None
-        Normalized category names, or None when no categories were provided.
-    """
-    if categories is None:
-        return None
-    return [_normalize_category(category) for category in categories]
-
-
-def _looks_like_curie(value: str) -> bool:
-    """
-    Return whether a string looks like a CURIE.
-
-    Parameters
-    ----------
-    value : str
-        Candidate node input.
-
-    Returns
-    -------
-    bool
-        True when the value appears to be a CURIE.
-    """
-    return ":" in value and " " not in value
-
-
-def _resolve_node(
-    value: str,
-    *,
-    name_resolver_kwargs: Optional[dict[str, Any]] = None,
-    node_normalizer_kwargs: Optional[dict[str, Any]] = None,
-) -> ResolvedNode:
-    """
-    Resolve a CURIE or display string into a normalized Translator node.
-
-    Parameters
-    ----------
-    value : str
-        CURIE or human-readable name.
-    name_resolver_kwargs : dict, optional
-        Additional keyword arguments for Name Resolver lookup.
-    node_normalizer_kwargs : dict, optional
-        Additional keyword arguments for Node Normalizer.
-
-    Returns
-    -------
-    ResolvedNode
-        Original input, preferred CURIE, label, and Biolink categories.
-
-    Raises
-    ------
-    LookupError
-        If the input cannot be resolved or normalized.
-    """
-    node_normalizer_kwargs = node_normalizer_kwargs or {}
-
-    if _looks_like_curie(value):
-        node = node_normalizer.get_normalized_nodes(value, **node_normalizer_kwargs)
-        if node is None:
-            raise LookupError(f"Could not normalize CURIE: {value}")
-    else:
-        resolved = name_resolver.lookup(
-            value,
-            return_top_response=True,
-            **(name_resolver_kwargs or {}),
-        )
-        node = (
-            node_normalizer.get_normalized_nodes(
-                resolved.curie,
-                **node_normalizer_kwargs,
-            )
-            or resolved
-        )
-
-    return ResolvedNode(
-        input_value=value,
-        curie=node.curie,
-        label=node.label,
-        categories=node.types or [],
-    )
-
-
-def _resolve_nodes(
-    values: Union[NodeInput, list[NodeInput]],
-    *,
-    name_resolver_kwargs: Optional[dict[str, Any]] = None,
-    node_normalizer_kwargs: Optional[dict[str, Any]] = None,
-) -> list[ResolvedNode]:
-    """
-    Resolve one or more node inputs.
-
-    Parameters
-    ----------
-    values : str or list[str]
-        CURIE or display-string node inputs.
-    name_resolver_kwargs : dict, optional
-        Additional keyword arguments for Name Resolver lookup.
-    node_normalizer_kwargs : dict, optional
-        Additional keyword arguments for Node Normalizer.
-
-    Returns
-    -------
-    list[ResolvedNode]
-        Resolved nodes in the same order as input.
-    """
-    if isinstance(values, str):
-        values = [values]
-    return [
-        _resolve_node(
-            value,
-            name_resolver_kwargs=name_resolver_kwargs,
-            node_normalizer_kwargs=node_normalizer_kwargs,
-        )
-        for value in values
-    ]
-
-
-def _get_resources(
-    *,
-    resources: Optional[TranslatorResources] = None,
-    api_names: Optional[dict[str, str]] = None,
-    meta_kg: Optional[pd.DataFrame] = None,
-    api_predicates: Optional[dict[str, list[str]]] = None,
-) -> TranslatorResources:
-    """
-    Merge explicit resource overrides with caller-provided or singleton resources.
-
-    Parameters
-    ----------
-    resources : TranslatorResources, optional
-        Complete resource object to use as the base.
-    api_names, meta_kg, api_predicates : optional
-        Partial overrides for the base resource object.
-
-    Returns
-    -------
-    TranslatorResources
-        Complete resource bundle for a query.
-    """
-    base = resources or get_translator_resources()
-    return TranslatorResources(
-        api_names=api_names if api_names is not None else base.api_names,
-        meta_kg=meta_kg if meta_kg is not None else base.meta_kg,
-        api_predicates=api_predicates
-        if api_predicates is not None
-        else base.api_predicates,
-    )
-
-
-def _build_finder_result(
-    raw_output: dict[str, Any],
-    *,
-    resolved_nodes: dict[str, ResolvedNode],
-) -> FinderResult:
-    """
-    Build a FinderResult from a parsed TRAPI-style output dictionary.
-
-    Parameters
-    ----------
-    raw_output : dict
-        Parsed output from an existing finder parser.
-    resolved_nodes : dict[str, ResolvedNode]
-        Resolved input-node metadata keyed by role.
-
-    Returns
-    -------
-    FinderResult
-        Convenience result wrapper.
-    """
-    return FinderResult(
-        query=raw_output.get("query_graph", {}),
-        knowledge_graph=raw_output.get("knowledge_graph", {}),
-        results=raw_output.get("results", []),
-        auxiliary_graphs=raw_output.get("auxiliary_graphs", {}),
-        resolved_nodes=resolved_nodes,
-        raw=raw_output,
-    )
